celery/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
bottle==0.12.19
celery==4.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
configparser==5.1.0
construct==2.5.3
coverage==6.0.2
cryptography==36.0.0
distlib==0.3.3
Django==3.2.8
django-appconf==1.0.5
django-compressor==2.4.1
django-cors-headers==3.10.0
django-haystack==3.1.1
django-ipware==4.0.0
django-mdeditor==0.1.18
django-ratelimit==3.0.1
django-redis==5.0.0
django-redis-cache==3.0.0
django-uuslug==1.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
filelock==3.0.12
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
haystack==0.42
idna==3.3
jieba==0.42.1
jsonpickle==2.0.0
Markdown==3.3.4
mysqlclient==2.0.3
pbr==5.6.0
pefile==2021.9.3
pika==1.2.0
Pillow==8.4.0
platformdirs==2.4.0
Pygments==2.10.0
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
python-logstash==0.4.6
python-memcached==1.59
python-ptrace==0.9.8
python-slugify==5.0.2
pytz==2021.3
raven==6.10.0
rcssmin==1.0.6
redis==3.5.3
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
stevedore==3.4.0
text-unidecode==1.3
ua-parser==0.10.0
urllib3==1.26.7
user-agents==2.2.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
WeRoBot==1.13.1
Whoosh==2.7.4
xmltodict==0.12.0
zope.event==4.5.0
zope.interface==5.4.0
